+++ arrow-libs/optics/arrow-optics/src/main/kotlin/arrow/optics/extensions/sequencek/index/SequenceKIndex.kt
@@ -23,7 +23,7 @@ internal val index_singleton: SequenceKIndex<Any?> = object : SequenceKIndex<Any
   "arrow.optics.extensions package is being deprecated. Use the exposed function in the instance for Sequence from the companion object of the typeclass instead.",
   ReplaceWith(
     "Index.sequence<A>().index(i)",
-    "arrow.optics.sequence", "arrow.optics.typeclasses.Index"
+    "arrow.optics.typeclasses.Index"
   ),
   DeprecationLevel.WARNING
 )
@@ -42,7 +42,7 @@ fun <A> index(i: Int): POptional<SequenceK<A>, SequenceK<A>, A, A> = arrow.core.
   "arrow.optics.extensions package is being deprecated, and it will be removed in 0.13.",
   ReplaceWith(
     "this compose Index.sequence<A>().index(i)",
-    "arrow.optics.compose", "arrow.optics.sequence", "arrow.optics.typeclasses.Index"
+    "arrow.optics.compose", "arrow.optics.typeclasses.Index"
   ),
   level = DeprecationLevel.WARNING
 )
@@ -59,7 +59,7 @@ operator fun <A, T> PLens<T, T, SequenceK<A>, SequenceK<A>>.get(i: Int): POption
   "SequenceK is being deprecated. Use the instance for Sequence from the companion object of the typeclass.",
   ReplaceWith(
     "Index.sequence<A>()",
-    "arrow.optics.sequence", "arrow.optics.typeclasses.Index"
+    "arrow.optics.typeclasses.Index"
   ),
   DeprecationLevel.WARNING
 )
