+++ arrow-libs/optics/arrow-optics/src/test/kotlin/arrow/optics/instances/SequenceInstanceTest.kt
@@ -8,7 +8,6 @@ import arrow.core.extensions.sequencek.eq.eq
 import arrow.core.test.UnitSpec
 import arrow.core.test.generators.functionAToB
 import arrow.optics.Traversal
-import arrow.optics.sequence
 import arrow.optics.test.laws.OptionalLaws
 import arrow.optics.test.laws.TraversalLaws
 import arrow.optics.typeclasses.FilterIndex
